refactor(catalog): extract exceptions into dedicated module
Move CatalogPushError and CatalogConfigurationError out of catalog.py
into exceptions.py, eliminating the lazy __getattr__ machinery in
__init__.py that existed only to avoid eagerly loading catalog.py for
these two lightweight classes.
